Update rbdnew03.lua
Extensive changes to how the script determines whether the player launched a day wrecker at the "Shield Control Tower" properly or not. It now checks the "who" of the Armory at launch time (valid for one frame!) to determine the target. If the target is incorrect, the mission immediately fails, if it is correct, the objective turns yellow initially and then green on impact, allowing the mission to continue.